introspection[asset-factory]: dump asset factories

+++ witch/common/src/engine/mod.rs
@@ -11,14 +11,14 @@ pub use r#unsafe::mutex::LuminousMutex;
 pub use r#unsafe::pointer::{LuminousCString, LuminousIntrusivePointer, LuminousPointer};
 pub use r#unsafe::string::LuminousString;
 
-pub const EBEX_OBJECT_ARRAY_ADDR_FORSPOKEN: usize = 0x77c5660;
-pub const EBEX_OBJECT_ARRAY_ADDR_XV: usize = 0x4f5bd30;
-pub const ASSET_FACTORY_CONTAINER_ADDR_FORSPOKEN: usize = 0x79d3458;
-pub const ASSET_FACTORY_CONTAINER_ADDR_XV: usize = 0x4ce7ba8;
-pub const ASSET_FACTORY_CONTAINER_OFFS_FORSPOKEN: usize = 0x5200;
-pub const ASSET_FACTORY_CONTAINER_OFFS_XV: usize = 0x8e0;
+pub const EBEX_OBJECT_ARRAY_ADDR_FORSPOKEN: u64 = 0x77c5660;
+pub const EBEX_OBJECT_ARRAY_ADDR_XV: u64 = 0x4f5bd30;
+pub const ASSET_FACTORY_CONTAINER_ADDR_FORSPOKEN: u64 = 0x79d3458;
+pub const ASSET_FACTORY_CONTAINER_ADDR_XV: u64 = 0x4ce7ba8;
+pub const ASSET_FACTORY_CONTAINER_OFFS_FORSPOKEN: u64 = 0x5200;
+pub const ASSET_FACTORY_CONTAINER_OFFS_XV: u64 = 0x8e0;
 
-#[derive(Debug, Eq, Hash, PartialEq)]
+#[derive(Debug, Copy, Clone, Eq, Hash, PartialEq)]
 pub enum LuminousGame {
 	LuminousEngine,
 	FinalFantasyXV,
